+++ jumi-core/src/main/java/fi/jumi/core/output/SynchronizedPrintStream.java
@@ -1,4 +1,4 @@
-// Copyright © 2011-2012, Esko Luontola <www.orfjackal.net>
+// Copyright © 2011-2013, Esko Luontola <www.orfjackal.net>
 // This software is released under the Apache License 2.0.
 // The license text is at http://www.apache.org/licenses/LICENSE-2.0
 
@@ -56,13 +56,16 @@ class SynchronizedPrintStream {
             this.lock = lock;
         }
 
+        @SuppressWarnings("SynchronizationOnLocalVariableOrMethodParameter")
         @Override
         public Object intercept(Object obj, Method method, Object[] args, MethodProxy proxy) throws Throwable {
-            lock.lock();
-            try {
-                return proxy.invokeSuper(obj, args);
-            } finally {
-                lock.unlock();
+            synchronized (obj) {
+                lock.lock();
+                try {
+                    return proxy.invokeSuper(obj, args);
+                } finally {
+                    lock.unlock();
+                }
             }
         }
     }