+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@ Release Notes
 
 - JUnit backward compatibilty: Jumi now runs JUnit 3, JUnit 4 and `@RunWith` annotated test classes. Requires you to have JUnit on your test classpath
 - Sets the context class loader for test threads ([Issue #2](https://github.com/orfjackal/jumi/issues/2))
+- Fixed a deadlock when calling `Throwable.printStackTrace()` and printing to standard output concurrently ([Issue #3](https://github.com/orfjackal/jumi/issues/3))
 
 ### Jumi 0.3.257 (2013-01-07)